Some changes

+++ frontend/src/app/page/restore-account/restore-account.component.ts
@@ -17,6 +17,8 @@ export class RestoreAccountComponent implements OnInit {
 
   email = '';
   token = '';
+  password1 = '';
+  password2 = '';
   lastTokenObject: { token: string; } = { token: '' };
   error = new BehaviorSubject({ hasError: false, message: '' });
 
@@ -42,7 +44,7 @@ export class RestoreAccountComponent implements OnInit {
 
   async checkTokens(): Promise<boolean> {
     if (!this.email || !this.token) {
-      this.error.next({ hasError: true, message: 'Hibás jelszóhelyreállítási link!' });
+      this.error.next({ hasError: true, message: 'Érvénytelen vagy lejárt jelszóhelyreállítási link!' });
       return false;
     }
 
@@ -59,33 +61,49 @@ export class RestoreAccountComponent implements OnInit {
       return false;
     }
   }
-  
+
   async setNewPassword(password: string): Promise<void> {
     const loginUrl = `${this.adminService.apiUrl}/login`;
     const logoutUrl = `${this.adminService.apiUrl}/logout`;
     const loginData = await this.http.post<any>(loginUrl, { email: 'rendszergazda@kft.hu', password: 'admin' }).toPromise();
     const admin = await this.adminService.findByEmail(this.email, loginData.accessToken, loginData._id).toPromise();
 
-    await this.adminService.update({ ...admin, password }, 
-      { 'headers': { 'Authorization': `Bearer ${loginData.accessToken}`, 'sessionId': loginData._id } 
-    }).toPromise();
+    await this.adminService.update({ ...admin, password },
+      {
+        'headers': { 'Authorization': `Bearer ${loginData.accessToken}`, 'sessionId': loginData._id }
+      }).toPromise();
 
     await this.http.post<{}>(logoutUrl, { sessionId: loginData._id, token: loginData.accessToken }).toPromise();
   }
-  
+
   async onSubmit(form: NgForm): Promise<void> {
-    try {
-      await this.setNewPassword(form.value.password1);
-      const deleteAllRequestByEmailUrl = `${this.adminService.apiUrl}/restoreAccount/?email=${this.email}`;
-      await this.http.delete<any>(deleteAllRequestByEmailUrl).toPromise();
-      await this.auth.logout(true);
-      this.toastr.success('Sikeresen módosítottad a jelszavad!', 'Siker!', {
-        timeOut: 5000,
-      });
-    } catch (error) {
-      this.toastr.error('Nem sikerült a jelszavad módosítása!', 'Hiba!', {
-        timeOut: 5000,
-      });
+    this.password1 = (document.querySelector('#password1Field') as HTMLInputElement).value;
+    this.password2 = (document.querySelector('#password2Field') as HTMLInputElement).value;
+    if (this.matchingPasswords()) {
+      try {
+        await this.setNewPassword(form.value.password1);
+        const deleteAllRequestByEmailUrl = `${this.adminService.apiUrl}/restoreAccount/?email=${this.email}`;
+        await this.http.delete<any>(deleteAllRequestByEmailUrl).toPromise();
+        await this.auth.logout(true);
+        this.toastr.success('Sikeresen módosult a jelszavad!', 'Siker!', {
+          timeOut: 5000,
+        });
+      } catch (error) {
+        this.toastr.error('Nem sikerült a jelszavad módosítása!', 'Hiba!', {
+          timeOut: 5000,
+        });
+      }
     }
   }
+
+  matchingPasswords(): boolean {
+    if (this.password1 !== this.password2) {
+      this.toastr.error('Nem egyeznek a jelszavak! Ellenőrizd!', 'Hiba!', {
+        timeOut: 4000,
+      });
+      return false;
+    }
+    return true;
+  }
+
 }
